Simplify how metric collector is hooked into OperationContext
